Removes bom and tool-list
These were kept for backwards compatibility with the IOP OKH schema, but since the Helpful Schema effectively _extends_ that schema, the fields are redundant. 

tool-list: is a plaintext list of tools

`tool-list: hammer, screwdriver`

However, this data is already contained within the `description` field of Helpful's extension, `tool-list-atoms`.

I proposed that in order to preserve backwards compatibility, Helpful's tooling should add an export to IOP specification which would extract the `description` field from the `tool-list-atoms`, and flatten that into the plaintext list that conforms to IOP spec. This is a simple operation.